requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.4.2
pydantic-settings==2.0.3
chromadb==0.4.15
openai==1.3.5
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
asyncio-throttle==1.0.2
tenacity==8.2.3
pytest==7.4.3
pytest-asyncio==0.21.1

# System Monitoring
psutil==5.9.6

# Vector Store
numpy<2.0.0

# File Processing Dependencies
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.3

# Optional file processing (will install only if available)
# For advanced PDF processing
# pymupdf==1.23.8
# For image processing
# Pillow==10.1.0
# For text extraction from various formats
# textract==1.6.5 